- fix `C_SetNotifyFontScale()` for RR.
Fixes #264.

+++ source/common/console/c_console.h
@@ -83,7 +83,7 @@ bool C_Responder (event_t *ev);
 void C_AddTabCommand (const char *name);
 void C_RemoveTabCommand (const char *name);
 void C_ClearTabCommands();		// Removes all tab commands
-void C_SetNotifyFontScale(int scale);
+void C_SetNotifyFontScale(double scale);
 
 extern const char *console_bar;
